Create a `SimpleReferenceRecorder` that can be used if the `GitReferenceRecorder` does not work.
