+++ src/main/resources/io/jenkins/plugins/forensics/reference/Messages.properties
@@ -1,2 +1,2 @@
-Recorder.DisplayName=Reference Build Finder
+Recorder.DisplayName=Discover reference build
 FieldValidator.Error.ReferenceJobDoesNotExist=There is no such job - maybe the job has been renamed?
