+++ src/main/resources/io/jenkins/plugins/forensics/reference/SimpleReferenceRecorder/config.jelly
@@ -1,0 +1,8 @@
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:f="/lib/form">
+
+  <f:entry title="${%title.referenceJob}" field="referenceJob" description="${%description.referenceJob}">
+    <f:combobox/>
+  </f:entry>
+
+</j:jelly>
