+++ src/main/resources/io/jenkins/plugins/forensics/reference/SimpleReferenceRecorder/config.properties
@@ -1,0 +1,2 @@
+title.referenceJob=Reference Job
+description.referenceJob=A reference job defines the baseline for tools that want to create relative build reports.
