+++ src/main/resources/io/jenkins/plugins/forensics/reference/SimpleReferenceRecorder/help-referenceJob.html
@@ -1,0 +1,3 @@
+The reference job is the baseline that is used to determine which of the issues in the current
+build are new, outstanding, or fixed. This baseline is also used to determine the number of new issues for the
+quality gate evaluation.
